Updated one step from then to step so it can be used more than once . Combined assigning case features into one feature file and py file. Removed function to create eligible finances and used already created steps.
