--- behave/features/assign_incomplete_case.feature
@@ -1,17 +1,0 @@
-
-Feature: Assign incomplete case.
-    Confirms that case notes are needed before assigning
-    a case.
-
-Background: Login
-    Given that I am logged in
-
-Scenario: Attempt to assign an incomplete case
-    Given I select to 'Create a case'
-    And case notes are empty
-    And I have created a user
-    And I have created a valid discrimination scope
-    And I am on the Diversity tab
-    When I select 'Prefer not say' for all diversity questions
-    And select the Assign tab
-    Then I get a message with the text "Case notes must be added to close a case"
